--- projects/concave/crv_abi.json
@@ -1,3 +1,0 @@
-{
-    "crvLP_coins": {"stateMutability":"view","type":"function","name":"coins","inputs":[{"name":"arg0","type":"uint256"}],"outputs":[{"name":"","type":"address"}],"gas":3123}
-}
