--- projects/concave/cvx_abi.json
@@ -1,9 +1,0 @@
-{
-    "cvxBRP_pid": {"inputs":[],"name":"pid","outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"stateMutability":"view","type":"function"}, 
-    "cvxBRP_balanceOf": {"inputs":[{"internalType":"address","name":"account","type":"address"}],"name":"balanceOf","outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"stateMutability":"view","type":"function"}, 
-    "cvxBRP_earned": {"inputs":[{"internalType":"address","name":"account","type":"address"}],"name":"earned","outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"stateMutability":"view","type":"function"}, 
-    "cvxBRP_rewards": {"inputs":[{"internalType":"address","name":"","type":"address"}],"name":"rewards","outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"stateMutability":"view","type":"function"}, 
-    "cvxBRP_userRewardPerTokenPaid": {"inputs":[{"internalType":"address","name":"","type":"address"}],"name":"userRewardPerTokenPaid","outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"stateMutability":"view","type":"function"}, 
-    "cvxBRP_stakingToken": {"inputs": [],"name": "stakingToken","outputs": [ { "internalType": "address", "name": "stakingToken", "type": "address" } ],"stateMutability": "view","type": "function"},
-    "cvxBooster_poolInfo": {"inputs": [{ "internalType": "uint256", "name": "", "type": "uint256" }],"name": "poolInfo","outputs": [ { "internalType": "address", "name": "lptoken", "type": "address" }, { "internalType": "address", "name": "token", "type": "address" }, { "internalType": "address", "name": "gauge", "type": "address" }, { "internalType": "address", "name": "crvRewards", "type": "address" }, { "internalType": "address", "name": "stash", "type": "address" }, { "internalType": "bool", "name": "shutdown", "type": "bool" } ],"stateMutability": "view","type": "function"}
-}
